Core: Updated to 11.2.0
Port From (https://github.com/TrinityCore/TrinityCore/commit/5cf0c6c8bb2c4e58a2d66ba5f304af34d18a4782)

+++ Source/Game/DataStorage/Structs/U_Records.cs
@@ -109,7 +109,7 @@ namespace Game.DataStorage
         public float RegenerationPeace;
         public float RegenerationCombat;
         public byte BarType;
-        public ushort Flags;
+        public int Flags;
         public float StartInset;
         public float EndInset;
         public uint[] FileDataID = new uint[6];
